Speed Pt.2 - Trackers Glory:
Made all the trackers load from disk on init, that way need need to publish less changes to swiftUI making their data available much earlier

Changes the startup sequence to load a community view much earlier
