+++ Mlem/MlemApp.swift
@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import UIKit
 
 @main
 struct MlemApp: App {
@@ -21,73 +22,25 @@ struct MlemApp: App {
                 .environmentObject(accountsTracker)
                 .onAppear {
                     URLCache.shared = AppConstants.urlCache
-                    loadAccounts()
 
                     setupAppShortcuts()
 
                     // set app theme to user preference
                     let windowScene =  UIApplication.shared.connectedScenes.first as? UIWindowScene
                     windowScene?.windows.first?.overrideUserInterfaceStyle = lightOrDarkMode
+
+                    let appearance = UITabBarAppearance()
+                    appearance.backgroundEffect = UIBlurEffect.init(style: .systemThinMaterial)
+                    UITabBar.appearance().standardAppearance = appearance
+                    UITabBar.appearance().scrollEdgeAppearance = appearance
                 }
         }
-        .onChange(of: accountsTracker.savedAccounts) { newValue in
-            do {
-                let encodedSavedAccounts: Data = try encodeForSaving(object: newValue)
-
-                do {
-                    try writeDataToFile(data: encodedSavedAccounts, fileURL: AppConstants.savedAccountsFilePath)
-                } catch let writingError {
-                    print("Failed while saving data to file: \(writingError)")
-                }
-            } catch let encodingError {
-                print("Failed while encoding accounts to data: \(encodingError)")
-            }
+        .onChange(of: accountsTracker.savedAccounts) { _ in
+            accountsTracker.saveToDisk()
         }
         .onChange(of: lightOrDarkMode) { value in
             let windowScene =  UIApplication.shared.connectedScenes.first as? UIWindowScene
             windowScene?.windows.first?.overrideUserInterfaceStyle = value
-        }
-    }
-
-    func loadAccounts() {
-        if FileManager.default.fileExists(atPath: AppConstants.savedAccountsFilePath.path) {
-            print("Saved Accounts file exists, will attempt to load saved accounts")
-            do {
-                let loadedUpAccounts = try decodeFromFile(
-                    fromURL: AppConstants.savedAccountsFilePath,
-                    whatToDecode: .accounts
-                ) as? [SavedAccount] ?? []
-
-                // MARK: - Associate the accounts with their secret credentials
-
-                if !loadedUpAccounts.isEmpty {
-
-                    let accounts = loadedUpAccounts.compactMap { account -> SavedAccount? in
-                        guard let token = AppConstants.keychain["\(account.id)_accessToken"] else {
-                            return nil
-                        }
-
-                        return SavedAccount(
-                            id: account.id,
-                            instanceLink: account.instanceLink,
-                            accessToken: token,
-                            username: account.username
-                        )
-                    }
-
-                    accountsTracker.savedAccounts = accounts
-                }
-            } catch let savedAccountDecodingError {
-                print("Failed while decoding saved accounts: \(savedAccountDecodingError)")
-            }
-        } else {
-            print("Saved Accounts file does not exist, will try to create it")
-
-            do {
-                try createEmptyFile(at: AppConstants.savedAccountsFilePath)
-            } catch let emptyFileCreationError {
-                print("Failed while creating an empty file: \(emptyFileCreationError)")
-            }
         }
     }
