+++ Mlem/Window.swift
@@ -12,54 +12,13 @@ struct Window: View {
     @StateObject var favoriteCommunitiesTracker: FavoriteCommunitiesTracker = .init()
     @StateObject var communitySearchResultsTracker: CommunitySearchResultsTracker = .init()
     @StateObject var filtersTracker: FiltersTracker = .init()
-    
+
     var body: some View {
         ContentView()
             .environmentObject(filtersTracker)
             .environmentObject(appState)
             .environmentObject(favoriteCommunitiesTracker)
             .environmentObject(communitySearchResultsTracker)
-            .onAppear {
-                if FileManager.default.fileExists(atPath: AppConstants.filteredKeywordsFilePath.path) {
-                    print("Filtered keywords file exists, will attempt to load blocked keywords")
-                    do {
-                        filtersTracker.filteredKeywords = try decodeFromFile(
-                            fromURL: AppConstants.filteredKeywordsFilePath,
-                            whatToDecode: .filteredKeywords
-                        ) as? [String] ?? []
-                    } catch let savedKeywordsDecodingError {
-                        print("Failed while decoding saved filtered keywords: \(savedKeywordsDecodingError)")
-                    }
-                } else {
-                    print("Filtered keywords file does not exist, will try to create it")
-                    
-                    do {
-                        try createEmptyFile(at: AppConstants.filteredKeywordsFilePath)
-                    } catch let emptyFileCreationError {
-                        print("Failed while creating an empty file: \(emptyFileCreationError)")
-                    }
-                }
-                print("now filtering: \(filtersTracker.filteredKeywords.count)")
-                if FileManager.default.fileExists(atPath: AppConstants.favoriteCommunitiesFilePath.path) {
-                    print("Favorite communities file exists, will attempt to load favorite communities")
-                    do {
-                        favoriteCommunitiesTracker.favoriteCommunities = try decodeFromFile(
-                            fromURL: AppConstants.favoriteCommunitiesFilePath,
-                            whatToDecode: .favoriteCommunities
-                        ) as? [FavoriteCommunity] ?? []
-                    } catch let favoriteCommunitiesDecodingError {
-                        print("Failed while decoding favorite communities: \(favoriteCommunitiesDecodingError)")
-                    }
-                } else {
-                    print("Favorite communities file does not exist, will try to create it")
-
-                    do {
-                        try createEmptyFile(at: AppConstants.favoriteCommunitiesFilePath)
-                    } catch let emptyFileCreationError {
-                        print("Failed while creating empty file: \(emptyFileCreationError)")
-                    }
-                }
-            }
             .onChange(of: filtersTracker.filteredKeywords) { newValue in
                 print("Change detected in filtered keywords: \(newValue)")
                 do {
